+++ ai-terminal/src-tauri/src/command/core/mod.rs
@@ -1,0 +1,1 @@
+pub mod execute_command;
