+++ ai-terminal/src-tauri/src/command/git_commands/mod.rs
@@ -1,0 +1,1 @@
+pub mod git;
