+++ ai-terminal/src-tauri/src/command/mod.rs
@@ -1,2 +1,4 @@
-pub mod types;
+pub mod core;
 pub mod git_commands;
+pub mod types;
+pub mod utils;
