+++ ai-terminal/src-tauri/src/command/types/command_state.rs
