+++ ai-terminal/src-tauri/src/command/types/mod.rs
