+++ ai-terminal/src-tauri/src/command/utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod path_utils;
