+++ ai-terminal/src-tauri/src/lib.rs
@@ -1,2 +1,3 @@
-pub mod ollama;
 pub mod command;
+pub mod ollama;
+pub mod utils;
