+++ ai-terminal/src-tauri/src/ollama/mod.rs
