+++ ai-terminal/src-tauri/src/ollama/types/mod.rs
