+++ ai-terminal/src-tauri/src/ollama/types/ollama_request.rs
