+++ ai-terminal/src-tauri/src/ollama/types/ollama_state.rs
